move a lot of stuff out of the astar header
